Introduce WITH_ASTEROIDAPP and WITH_CMAKE_MODULES
WITH_ASTEROIDAPP
The AsteroidApp class does not always need to be built, so this allows the
user to configure the build so that it is not.  It's ON by default for
backwards compatibility.

WITH_CMAKE_MODULES
The CMake module installation is similar in that it does not necessarily
need to be done.  If this is OFF the CMake modules provided by this
project will not be installed and the QML install dir will be determined
by using QMake to ask for the location.

Signed-off-by: Ed Beroset <beroset@ieee.org>