Redesign PageHeader component
- Add gradient from near black to transparent behind the title Label to mitigate clipping with of scrolling page content
- Change Fonttype to SemiCondensed Light, to be consistent with the Settings main ListItem
- Add wrapper Item around the rectangle and Label. This makes it necessary to forward the alias for text to the Label.text since some calls of the component assume it is a Label and push the title to the Label.text property.

This requires changes to all pages and apps where PageHeader is used. The PageHeader call needs to be moved last in code order to prevent the text content to scroll on top of the PageHeader.
Implementation of this change into the settings pages here https://github.com/AsteroidOS/asteroid-settings/pull/84

Signed-off-by: Timo Könnecke koennecke@mosushi.de