Remove width from IntSelector label
The recent refactoring removed the former labelWidthRatio property, but
the code was still referring to it to set the width of the label in the
middle of the IntSelector control.  This fix simply removes the explict
width setting because it is implied by the anchors and the widths of
the left and right buttons.

Signed-off-by: Ed Beroset <beroset@ieee.org>